remove 'old' BlogPageContainer and PostContainer, move AboutContainer
